[ops] move !name into ops/name.py
